Porting v0.8.1: fixed unintended integer math in some buff icon fading logic

+++ core/src/main/java/com/shatteredpixel/yasd/general/items/wands/WandOfLivingEarth.java
@@ -261,7 +261,7 @@ public class WandOfLivingEarth extends DamageWand {
 
 		@Override
 		public float iconFadePercent() {
-			return Math.max(0, (armorToGuardian() - armor) / armorToGuardian());
+			return Math.max(0, (armorToGuardian() - armor) / (float)armorToGuardian());
 		}
 
 		@Override
